rename Competition to Problems and move to components/
