+++ pages/matik/zadania/[series].tsx
@@ -1,7 +1,7 @@
 import {NextPage} from 'next'
 
 import {PageLayout} from '../../../src/components/PageLayout/PageLayout'
-import {Problems} from '../../../src/pages/Competition/Problems'
+import {Problems} from '../../../src/components/Problems/Problems'
 
 const ProblemsPage: NextPage = () => (
   <PageLayout seminarId={2}>
